chore: Update image URLs in client and server code

+++ server/controllers/foodController.js
@@ -1,5 +1,6 @@
 import foodModel from "../models/foodModel.js";
-import fs from "fs"; 
+import fs from "fs";
+import cloudinary from "cloudinary";
 
 // add food item
 
@@ -7,13 +8,15 @@ const addFood = async (req, res) => {
   if (!req.file) {
     return res.status(400).json({ success: false, message: "No file uploaded" });
   }
-  let image_filename = req.file.path;
+  let image_url = req.file.path;
+  let image_id = req.file.filename; 
 
   const food = new foodModel({
     name: req.body.name,
     description: req.body.description,
     price: req.body.price,
-    image: image_filename,
+    image: image_url,
+    image_id: image_id,
     category: req.body.category
   });
 
@@ -44,7 +47,11 @@ const listFood = async (req, res) => {
 const removeFood = async (req, res) => {
   try {
     const food = await foodModel.findById(req.body.id);
-    fs.unlink(`uploads/${food.image}`, () => {})
+    // Remove the image from Cloudinary
+    const result = await cloudinary.uploader.destroy(food.image_id); // Using Cloudinary's destroy method
+    if (result.result !== "ok") {
+      return res.status(500).json({ success: false, message: "Failed to delete image from Cloudinary" });
+    }
     await foodModel.findByIdAndDelete(req.body.id);
     res.json({success: true, message: "Food item removed successfully"});
   } catch (error) {
